// types.ts
export interface Video {
    id: string;
    songId: string;
    title: string;
    type: 'Music Video' | 'Lyric Video' | 'Visualizer' | 'Genius Verified' | 'Live Performance' | 'Interview' | 'Custom';
    views: number;
    thumbnail: string;
    releaseDate: GameDate;
    artistId: string;
    channelId: string;
    firstWeekViews?: number;
    description?: string;
    mentionedNpcs?: string[];
}
// ...
export interface GameDate {
    week: number;
    year: number;
}